IA/requirements-deploy-flask.txt
numpy
keras
tensorflow
pillow
pandas
matplotlib
Flask
opencv-python
IPython